Fix some types (please check if this fixes #1854), remove strnlen function from gbox module as it is not crossplattform and furthermore not necessary here. Add missing initialization for sc8in1 lock.
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@5233 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ module-cccshare.c
@@ -372,7 +372,7 @@ int32_t card_valid_for_client(struct s_client *cl, struct cc_card *card) {
         if (ll_count(card->providers)) {
         	it = ll_iter_create(card->providers);
         	struct cc_provider *prov;
-        	int found=0;
+        	int8_t found=0;
         	while ((prov = ll_iter_next(&it))) {
         		uint32_t prid = prov->prov;
                 if (chk_srvid_by_caid_prov(cl, card->caid, prid)) {
